Consolidate `ring-compat` from `ring-aead` and `ring-digest`
Merges the `ring-aead` and `ring-digest` crates together, producing a
unified `ring-compat` crate.
