--- .github/workflows/ring-aead.yml
@@ -1,58 +1,0 @@
-name: ring-aead
-
-on:
-  pull_request:
-    paths:
-      - "ring-aead/**"
-      - "Cargo.*"
-  push:
-    branches: master
-
-defaults:
-  run:
-    working-directory: ring-aead
-
-env:
-  CARGO_INCREMENTAL: 0
-  RUSTFLAGS: "-Dwarnings"
-
-jobs:
-  #  # TODO: test no_std builds when rust-lang/cargo#7916 is on stable
-  #  build:
-  #    runs-on: ubuntu-latest
-  #    strategy:
-  #      matrix:
-  #        rust:
-  #          - 1.41.0 # MSRV
-  #          - stable
-  #        target:
-  #          # - armv7-none-eabi TODO: test after MSRV is bumped to 1.42
-  #          - thumbv7em-none-eabi
-  #          - wasm32-unknown-unknown
-  #    steps:
-  #      - uses: actions/checkout@v1
-  #      - uses: actions-rs/toolchain@v1
-  #        with:
-  #          profile: minimal
-  #          toolchain: ${{ matrix.rust }}
-  #          target: ${{ matrix.target }}
-  #          override: true
-  #      - run: cargo build --no-default-features --release --target ${{ matrix.target }}
-
-  test:
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        rust:
-          - 1.41.0 # MSRV
-          - stable
-    steps:
-      - uses: actions/checkout@v1
-      - uses: actions-rs/toolchain@v1
-        with:
-          profile: minimal
-          toolchain: ${{ matrix.rust }}
-          override: true
-      - run: cargo test --lib --release
-      - run: cargo test --all-features --release
-
