+++ .github/workflows/ring-compat.yml
@@ -1,0 +1,29 @@
+name: ring-compat
+
+on:
+  pull_request: {}
+  push:
+    branches: master
+
+env:
+  CARGO_INCREMENTAL: 0
+  RUSTFLAGS: "-Dwarnings"
+
+jobs:
+  test:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        rust:
+          - 1.41.0 # MSRV
+          - stable
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions-rs/toolchain@v1
+        with:
+          profile: minimal
+          toolchain: ${{ matrix.rust }}
+          override: true
+      - run: cargo test --lib --release
+      - run: cargo test --all-features --release
+
