+++ CHANGELOG.md
@@ -3,10 +3,3 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## 0.2.0 (2020-06-13)
-- Update `aead` to v0.3.0.
-- Update `Cargo.toml` repository link.
-
-## 0.1.0 (2019-12-06)
-- Initial release
