--- ring-digest/README.md
@@ -1,38 +1,0 @@
-# RustCrypto: *ring* `Digest` trait wrappers
-
-[![crate][crate-image]][crate-link]
-[![Docs][docs-image]][docs-link]
-![Apache2/MIT licensed][license-image]
-![Rust Version][rustc-image]
-
-Wappers for using the high-performance implementations of digest algorithm from the *ring* crate via [`digest::Digest`] trait from the `digest` crate..
-
-[Documentation][docs-link]
-
-## License
-
-Licensed under either of:
-
- * [Apache License, Version 2.0](http://www.apache.org/licenses/LICENSE-2.0)
- * [MIT license](http://opensource.org/licenses/MIT)
-
-at your option.
-
-### Contribution
-
-Unless you explicitly state otherwise, any contribution intentionally submitted
-for inclusion in the work by you, as defined in the Apache-2.0 license, shall be
-dual licensed as above, without any additional terms or conditions.
-
-[//]: # (badges)
-
-[crate-image]: https://img.shields.io/crates/v/ring-digest.svg
-[crate-link]: https://crates.io/crates/ring-digest
-[docs-image]: https://docs.rs/ring-digest/badge.svg
-[docs-link]: https://docs.rs/ring-digest/
-[license-image]: https://img.shields.io/badge/license-Apache2.0/MIT-blue.svg
-[rustc-image]: https://img.shields.io/badge/rustc-1.41+-blue.svg
-
-[//]: # (general links)
-
-[*ring*]: https://github.com/briansmith/ring
