+++ src/lib.rs
@@ -6,14 +6,20 @@
 #![warn(missing_docs, rust_2018_idioms)]
 
 pub use aead;
+pub use digest::{self, Digest};
 
 use aead::{
     consts::{U0, U12, U16, U32},
-    generic_array::GenericArray,
     AeadInPlace, Buffer, Error, NewAead,
+};
+use core::mem;
+use digest::{
+    generic_array::{typenum::*, GenericArray},
+    BlockInput, FixedOutput, Reset, Update,
 };
 use ring::aead::LessSafeKey as Key; // (^～^;)ゞ
 use ring::aead::{Aad, Nonce, UnboundKey, AES_128_GCM, AES_256_GCM, CHACHA20_POLY1305};
+use ring::digest::Context;
 use zeroize::Zeroize;
 
 /// Authentication tags
@@ -136,3 +142,94 @@ impl Cipher {
         Ok(())
     }
 }
+
+macro_rules! impl_digest {
+    (
+        $(#[doc = $doc:tt])*
+        $name:ident, $hasher:ident, $block_len:ty, $output_size:ty
+    ) => {
+        $(#[doc = $doc])*
+        #[repr(transparent)]
+        #[derive(Clone)]
+        pub struct $name(Context);
+
+        impl $name {
+            fn take(&mut self) -> Context {
+                mem::replace(&mut self.0, Context::new(&ring::digest::$hasher))
+            }
+        }
+
+        impl Default for $name {
+            fn default() -> Self {
+                $name(Context::new(&ring::digest::$hasher))
+            }
+        }
+
+        impl Update for $name {
+            fn update(&mut self, data: impl AsRef<[u8]>) {
+                self.0.update(data.as_ref())
+            }
+        }
+
+        impl BlockInput for $name {
+            type BlockSize = $block_len;
+        }
+
+        impl FixedOutput for $name {
+            type OutputSize = $output_size;
+
+            fn finalize_into(self, out: &mut GenericArray<u8, Self::OutputSize>) {
+                *out = GenericArray::clone_from_slice(self.0.finish().as_ref());
+            }
+
+            fn finalize_into_reset(&mut self, out: &mut GenericArray<u8, Self::OutputSize>) {
+                *out = GenericArray::clone_from_slice(self.take().finish().as_ref());
+            }
+        }
+
+        impl Reset for $name {
+            fn reset(&mut self) {
+                mem::drop(self.take());
+            }
+        }
+
+        digest::impl_write!($name);
+        opaque_debug::implement!($name);
+    };
+}
+
+impl_digest!(
+    /// Structure representing the state of a SHA-1 computation
+    Sha1,
+    SHA1_FOR_LEGACY_USE_ONLY,
+    U64,
+    U20
+);
+impl_digest!(
+    /// Structure representing the state of a SHA-256 computation
+    Sha256,
+    SHA256,
+    U64,
+    U32
+);
+impl_digest!(
+    /// Structure representing the state of a SHA-384 computation
+    Sha384,
+    SHA384,
+    U128,
+    U48
+);
+impl_digest!(
+    /// Structure representing the state of a SHA-512 computation
+    Sha512,
+    SHA512,
+    U128,
+    U64
+);
+impl_digest!(
+    /// Structure representing the state of a SHA-512/256 computation
+    Sha512Trunc256,
+    SHA512_256,
+    U128,
+    U32
+);
