+++ tests/digest.rs
@@ -1,0 +1,1 @@
+#![no_std]
